// src/types/CellSprite.ts
export enum CellSprite {
    Close = "close",
    Selected = "selected",
    Bomb = "bomb",
    Flag = "flag",
    QuestionMark = "questionMark",
    ExplodedBomb = "explodedBomb",
    FalseFlag = "falseFlag",
    One = 1,
    Two = 2,
    Three = 3,
    Four = 4,
    Five = 5,
    Six = 6,
    Seven = 7,
    Eight = 8
  }